requirements.txt
pandas>=2.0.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
pydantic==2.9.2
pydantic-settings==2.6.1
psycopg[binary]==3.2.3
pyodbc==5.2.0
firebird-driver==1.10.6

